+++ apps/client/src/plugins/Print/PrintModel.js
@@ -20,7 +20,7 @@ import ImageWMS from "ol/source/ImageWMS";
 
 import QRCode from "qrcode";
 
-import { ROBOTO_NORMAL } from "./constants";
+import { ROBOTO_NORMAL, ROBOTO_BOLD } from "./constants";
 
 const DEFAULT_DIMS = {
   a0: [1189, 841],
@@ -44,6 +44,8 @@ export default class PrintModel {
     this.scaleMeters = settings.options.scaleMeters;
     this.scaleBarLengths = this.calculateScaleBarLengths();
     this.copyright = settings.options.copyright || "";
+    this.textFontSize = settings.options.textFontSize || 8;
+    this.textFontWeight = settings.options.textFontWeight || "normal";
     this.date = settings.options.date || "";
     this.disclaimer = settings.options.disclaimer || "";
     this.localObserver = settings.localObserver;
@@ -1130,17 +1132,24 @@ export default class PrintModel {
 
   // Adds fonts needed to properly render necessary characters. (The default jsPDF fonts does not support all characters).
   // Also enables a font (in the future we could provide a possibility for the user to select font).
-  setupFonts = (pdf, font = "ROBOTO_NORMAL") => {
+  setupFonts = (pdf, font) => {
     // First we'll add the available fonts
+    // Normal
     pdf.addFileToVFS("roboto-normal.ttf", ROBOTO_NORMAL);
     pdf.addFont("roboto-normal.ttf", "roboto-normal", "normal");
-    // Then we'll set the font we want to use now. (The switch below is unnecessary but
-    // added for possible future use cases).
+    // Bold
+    pdf.addFileToVFS("roboto-bold.ttf", ROBOTO_BOLD);
+    pdf.addFont("roboto-bold.ttf", "roboto-bold", "bold");
+    // Then we'll set the font that we want to use.
     switch (font) {
-      case "ROBOTO_NORMAL":
-        pdf.setFont("roboto-normal");
+      case "normal":
+        pdf.setFont("roboto-normal", "normal");
+        break;
+      case "bold":
+        pdf.setFont("roboto-bold", "bold");
         break;
       default:
+        pdf.setFont("roboto-normal", "normal");
         break;
     }
   };
@@ -1238,7 +1247,7 @@ export default class PrintModel {
         });
 
         // Make sure to add necessary fonts and enable the font we want to use.
-        this.setupFonts(pdf, "ROBOTO_NORMAL");
+        this.setupFonts(pdf, this.textFontWeight);
 
         // Add our map canvas to the PDF, start at x/y=0/0 and stretch for entire width/height of the canvas
         pdf.addImage(mapCanvas, "JPEG", 0, 0, dim[0], dim[1]);
@@ -1438,17 +1447,30 @@ export default class PrintModel {
             align: "center",
           });
         }
-
-        // Add potential copyright text
+        // Calculate the y-offset for the copyright and disclaimer text based on font size.
+        const yOffset =
+          this.textFontSize === 8
+            ? { copyright: 5.5, disclaimer: 6 }
+            : this.textFontSize === 11
+              ? { copyright: 7, disclaimer: 7.5 }
+              : this.textFontSize === 13
+                ? { copyright: 8, disclaimer: 9 }
+                : { copyright: 5.5, disclaimer: 6 };
+        //  Add potential copyright text
         if (this.copyright.length > 0) {
           let yPos = options.useTextIconsInMargin
             ? this.textIconsMargin + this.margin / 2
             : this.margin;
-          pdf.setFontSize(8);
+          pdf.setFontSize(this.textFontSize);
           pdf.setTextColor(options.mapTextColor);
-          pdf.text(this.copyright, dim[0] - 4 - yPos, dim[1] - 5.5 - yPos, {
-            align: "right",
-          });
+          pdf.text(
+            this.copyright,
+            dim[0] - 4 - yPos,
+            dim[1] - yOffset.copyright - yPos,
+            {
+              align: "right",
+            }
+          );
         }
 
         // Add potential date text
@@ -1460,7 +1482,7 @@ export default class PrintModel {
           let yPos = options.useTextIconsInMargin
             ? this.textIconsMargin + this.margin / 2
             : this.margin;
-          pdf.setFontSize(8);
+          pdf.setFontSize(this.textFontSize);
           pdf.setTextColor(options.mapTextColor);
           pdf.text(date, dim[0] - 4 - yPos, dim[1] - 2 - yPos, {
             align: "right",
@@ -1472,17 +1494,18 @@ export default class PrintModel {
           let yPos = options.useTextIconsInMargin
             ? this.textIconsMargin + this.margin / 2
             : this.margin;
-          pdf.setFontSize(8);
           pdf.setTextColor(options.mapTextColor);
           let textLines = pdf.splitTextToSize(
             this.disclaimer,
             dim[0] / 2 - this.margin - 8
           );
-          let textLinesDims = pdf.getTextDimensions(textLines, { fontSize: 8 });
+          let textLinesDims = pdf.getTextDimensions(textLines, {
+            fontSize: this.textFontSize,
+          });
           pdf.text(
             textLines,
             dim[0] - 4 - yPos,
-            dim[1] - 6 - yPos - textLinesDims.h,
+            dim[1] - yOffset.disclaimer - yPos - textLinesDims.h,
             {
               align: "right",
             }